Found the source of the backprop error
It has been haunting me for a month, but I have finally found the source
of the darkness. Like a ghost, hiding inconspicuously, where it should
be impossible for an error to occur.

But at least now I know where it's coming from now. Now just to figure
out why.